rc: opt-in GC on object creation
The previous ideas are inspired by bacon-rajan-cc: opt-in GC on drop, or on
clone. That actually implies the GC list can be incomplete and GC traversal
needs to be recursive, which in turn requires extra book-keeping to skip
visiting visited objects. I figured that the cpython "complete gc list"
approach is easier. So this crate is implemented the cpython way.

Here are some more detailed differences between cpython's gc, and
bacon-rajan-cc in this area:

GC list:
- cpython used linked list. bacon-rajan-cc uses a separate vec. cpython can
  drop object from the gc list when refcount reaches 0, while bacon-rajan-cc
  cannot.
- cpython's linked list contains all objects, bacon-rajan-cc's list only
  needs "roots". This affects "GC traverse".

GC traverse:
- cpython's tp_traverse only visits one level of objects because its gc list
  is complete. bacon-rajan-cc's visit is recursive since its gc list is
  incomplete.  bacon-rajan-cc requires extra bookkeeping to skip visited
  objects.

GC opt-in:
- cpython adds object to gc list on object creation. bacon-rajan-cc adds
  object on drop. Again, this is because cpython's gc list is complete and
  tp_traverse is not recursive.